Some clean-up. Add missing imports. Use spaces throughout the code.
