+++ django/account/api_0_1.py
@@ -27,11 +27,11 @@ from ai.api import ActiveAndAuthenticatedPermission
 from . import serializers
 
 class AuthView_0_1(APIView):
-	'''
-	AuthView returns the requesting User's info or an empty map if the request is not authed and active.
-	'''
-	permission_classes = (NoPermission,)
+    '''
+    AuthView returns the requesting User's info or an empty map if the request is not authed and active.
+    '''
+    permission_classes = (NoPermission,)
 
-	def get(self, request):
-		if request.user.is_anonymous() or not request.user.is_active: return Response({})
-		return Response(serializers.SelfUserSerializer_0_1(request.user, context={'request':request}).data)
+    def get(self, request):
+        if request.user.is_anonymous() or not request.user.is_active: return Response({})
+        return Response(serializers.SelfUserSerializer_0_1(request.user, context={'request':request}).data)
