+++ django/ai/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
 class StyledModelAdmin(admin.ModelAdmin):
-	"""A common base admin class with shared media information."""
-	class Media:
-		css = { "all": ('admin.css', )}
+    """A common base admin class with shared media information."""
+    class Media:
+        css = { "all": ('admin.css', )}
