+++ django/ai/local_settings.py.example
@@ -12,7 +12,7 @@ DATABASES = {
         'ENGINE': 'django.db.backends.postgresql_psycopg2',
         'NAME': 'ai',
         'USER': 'postgres',
-        'password': 'moo',
+        'PASSWORD': 'moo',
         'HOST': '127.0.0.1',
         'PORT': '5432'
     }
