+++ django/front/views.py
@@ -2,4 +2,4 @@ from django.shortcuts import render
 from django.contrib.auth.decorators import login_required
 
 def index(request):
-	return render(request, 'front/index.html', {})
+    return render(request, 'front/index.html', {})
