+++ django/heartbeat/urls.py
@@ -2,8 +2,8 @@ from django.contrib import admin
 from django.conf.urls import url
 from django.conf.urls import include
 
-import heartbeat.views as heartbeat_views 
+import heartbeat.views as heartbeat_views
 
 urlpatterns = [
-	url(r'^$', heartbeat_views.index, name='index'),
+    url(r'^$', heartbeat_views.index, name='index'),
 ]
