+++ django/requirements.txt
@@ -1,6 +1,6 @@
-Django==1.10.6
+Django==1.11.11
 Pillow==4.0.0
-djangorestframework==3.6.2
+djangorestframework==3.8.2
 gunicorn==19.7.1
 psycopg2==2.7.1
 pytz==2016.10
