+++ django/weather/management/commands/import_airports.py
@@ -17,38 +17,44 @@ from django.core.management.base import BaseCommand, CommandError
 
 from weather.models import Location
 
-class Command(BaseCommand):
-	help = "Imports info from the airports.csv into Location records."
-	requires_system_checks = True
 
-	def handle(self, *labels, **options):
-		airports_csv_path = os.path.join(settings.BASE_DIR, 'weather', 'airports.csv')
-		if not os.path.isfile(airports_csv_path):
-			print 'Can not find %s' % airports_csv_path
-			return
-		airport_reader = csv.reader(open(airports_csv_path))
-		count = 0
-		Location.objects.all().delete()
-		for row in airport_reader:
-			Location.objects.create(
-				name=row[0],
-				airport_type=row[1],
-				latitude=parse_float_field(row[2]),
-				longitude=parse_float_field(row[3]),
-				elevation=parse_float_field(row[4]),
-				iso_country=row[5],
-				iso_region=row[6],
-				municipality=row[7],
-				gps_code=row[8],
-				iata_code=row[9],
-				local_code=row[10]
-			)
-			count +=1 
-			if count % 1000 == 0:
-				sys.stdout.write('.')
-				sys.stdout.flush()
-		sys.stdout.write('\n')
+class Command(BaseCommand):
+
+    help = "Imports info from the airports.csv into Location records."
+
+    def handle(self, *labels, **options):
+
+        airports_csv_path = os.path.join(settings.BASE_DIR, 'weather', 'airports.csv')
+
+        if not os.path.isfile(airports_csv_path):
+            print 'Can not find %s' % airports_csv_path
+            return
+
+        airport_reader = csv.reader(open(airports_csv_path))
+        count = 0
+        Location.objects.all().delete()
+
+        for row in airport_reader:
+            Location.objects.create(
+                name=row[0],
+                airport_type=row[1],
+                latitude=parse_float_field(row[2]),
+                longitude=parse_float_field(row[3]),
+                elevation=parse_float_field(row[4]),
+                iso_country=row[5],
+                iso_region=row[6],
+                municipality=row[7],
+                gps_code=row[8],
+                iata_code=row[9],
+                local_code=row[10]
+            )
+            count +=1
+            if count % 1000 == 0:
+                sys.stdout.write('.')
+                sys.stdout.flush()
+        sys.stdout.write('\n')
+
 
 def parse_float_field(val):
-	if val == '': return None
-	return float(val)
+    if val == '': return None
+    return float(val)
